// api/generated/models/createUsuarioDtoNivel.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Autenticación: POST /auth/login fija la cookie HttpOnly access_token. En Swagger UI, ejecuta login en este mismo origen y luego las rutas protegidas enviarán la cookie automáticamente. Desde el frontend usa fetch/axios con credentials: include.
 * OpenAPI spec version: 1.0
 */

export type CreateUsuarioDtoNivel = typeof CreateUsuarioDtoNivel[keyof typeof CreateUsuarioDtoNivel];


export const CreateUsuarioDtoNivel = {
  USUARIO: 'USUARIO',
  ADMINISTRADOR: 'ADMINISTRADOR',
} as const;